requirements.txt
opencv-python
pytesseract
transformers
torch  # Required for transformers to run summaries
Pillow
